Fix the Xcode project, split the messageDepthTest off.
- Fix SwiftProtobuf.xcodeproj by adding JSONDecodingOptions.swift to it.
- Move the messageDepthLimit test out into a new file since the file it was
  added to is supposed to be extracted conformance test tests.
